+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 <!-- _No changes yet._ -->
 
 - feat: add a scale effect while clicking on interactive widgets
+- feat: add bar element background opacity setting and apply it across widgets
+- feat: enable border display in global settings by default
+- feat: add showElementsBorder setting and apply it across widgets
 
 ## v1.3.3 - 2026-02-13